refactor: split editor config and link bubble modules
